pyproject.toml
[project]
name = "yonote-mcp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fast-agent-mcp>=0.2.23",
    "requests>=2.32.3",
]
